+++ neuron.h
@@ -9,26 +9,26 @@ class Neuron {
 private :
 	double V_; //membrane potential (mV)
 	int nb_spikes_;
-	std::vector<int> spike_times_;
+	std::vector<double> spike_times_;
 	double threshold_; //potential for which the neuron spikes (mV)
 	double tau_; //period where neuron is refractory (seconds)
-	int time_;
+	double time_;
 	bool is_refractory_;
 	double ref_period_;
 
 public :
 	
 	//Constructor
-	Neuron(int t);
+	Neuron(double t);
 	
 	//getters and setters
 	double getV();
 	int getNb_spikes();
-	std::vector<int> getSpike_times();
-	int getTime();
+	std::vector<double> getSpike_times();
+	double getTime();
 	void setV(double);
 	void setNb_spikes(int);
-	void setTime (int);
+	void setTime (double);
 	
 	//update method
 	void update(double);
